Changes after using more powerful openai model
